Rename package, move the dependency
