+++ modules/compute/src/jobs/java/org/apache/ignite/internal/compute/ConcatJob.java
@@ -15,10 +15,19 @@
  * limitations under the License.
  */
 
-package org.example;
+package org.apache.ignite.internal.compute;
 
-class JobException extends RuntimeException {
-    JobException(String message, Throwable cause) {
-        super(message, cause);
+import java.util.Arrays;
+import java.util.stream.Collectors;
+import org.apache.ignite.compute.ComputeJob;
+import org.apache.ignite.compute.JobExecutionContext;
+
+/** Compute job that concatenates the string representation of its arguments. */
+public class ConcatJob implements ComputeJob<String> {
+    @Override
+    public String execute(JobExecutionContext context, Object... args) {
+        return Arrays.stream(args)
+                .map(Object::toString)
+                .collect(Collectors.joining());
     }
 }
