+++ modules/compute/src/jobs/java/org/apache/ignite/internal/compute/FailingJob.java
@@ -15,21 +15,15 @@
  * limitations under the License.
  */
 
-package org.example;
+package org.apache.ignite.internal.compute;
 
-import java.util.concurrent.TimeUnit;
 import org.apache.ignite.compute.ComputeJob;
 import org.apache.ignite.compute.JobExecutionContext;
 
-/** Compute job that sleeps for a number of milliseconds passed in the argument. */
-public class SleepJob implements ComputeJob<Void> {
+/** Compute job that always fails with the {@link JobException}. */
+public class FailingJob implements ComputeJob<String> {
     @Override
-    public Void execute(JobExecutionContext jobExecutionContext, Object... args) {
-        try {
-            TimeUnit.SECONDS.sleep((Long) args[0]);
-            return null;
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
+    public String execute(JobExecutionContext context, Object... args) {
+        throw new JobException("Oops", new Exception());
     }
 }
