+++ modules/compute/src/unit1/java/org/apache/ignite/internal/compute/UnitJob.java
@@ -15,15 +15,15 @@
  * limitations under the License.
  */
 
-package org.example;
+package org.apache.ignite.internal.compute;
 
 import org.apache.ignite.compute.ComputeJob;
 import org.apache.ignite.compute.JobExecutionContext;
 
-/** Compute job that always fails with the {@link JobException}. */
-public class FailingJob implements ComputeJob<String> {
+/** Compute job. */
+public class UnitJob implements ComputeJob<Integer> {
     @Override
-    public String execute(JobExecutionContext context, Object... args) {
-        throw new JobException("Oops", new Exception());
+    public Integer execute(JobExecutionContext context, Object... args) {
+        return 1;
     }
 }
